requirements.txt
Pillow
requests
RPi.GPIO
